+++ app/src/androidTest/java/com/ing/software/ticketapp/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.ing.software.appscontrini;
+package com.ing.software.ticketapp;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.ing.software.appscontrini", appContext.getPackageName());
+        assertEquals("com.ing.software.ticketapp", appContext.getPackageName());
     }
 }
