+++ app/src/main/java/com/ing/software/ticketapp/OCR/OcrHandler.java
@@ -1,4 +1,4 @@
-package com.ing.software.appscontrini.OCR;
+package com.ing.software.ticketapp.OCR;
 
 import android.app.Service;
 import android.content.Intent;
