+++ app/src/main/java/com/ing/software/ticketapp/OCR/RawObjects.java
@@ -1,4 +1,4 @@
-package com.ing.software.appscontrini.OCR;
+package com.ing.software.ticketapp.OCR;
 
 import android.graphics.Bitmap;
 import android.graphics.RectF;
@@ -9,7 +9,6 @@ import com.google.android.gms.vision.text.TextBlock;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Scanner;
 
 /**
  * Class to store objects detected.
@@ -67,12 +66,14 @@ class RawBlock {
         return rawTexts;
     }
 
+
     /**
      * Loops throw Rawtexts checking if their rect is in a box where probability
      * to find amount is > 0
      * @param level number of results to ignore (used for deeper analysis)
      * @return string with detected amount, null if nothing is found
      */
+    /*
     String findAmount(int level) {
         String amount = null;
         int i = 0;
@@ -87,6 +88,7 @@ class RawBlock {
         }
         return null;
     }
+    */
 
     /**
      * Search string in block, only first occurrence is returned (top -> bottom, left -> right)
@@ -207,6 +209,7 @@ class RawBlock {
          * checks if amount is present
          * @return string with detected amount, null if nothing found
          */
+        /*
         private String findAmount() {
             int[] gridBox = getGridBox();
             int probability = ProbGrid.amountMap.get(grid)[gridBox[1]][gridBox[0]];
@@ -218,11 +221,13 @@ class RawBlock {
             else
                 return null;
         }
+        */
 
         /**
          * Find box of the grid containing the center of the text rect
          * @return coordinates of the grid, where int[0] = column, int[1] = row
          */
+        /*
         private int[] getGridBox() {
             Scanner gridder = new Scanner(grid);
             gridder.useDelimiter("x");
@@ -235,11 +240,13 @@ class RawBlock {
             int gridY = (int) (rectText.centerY()/rowsHeight);
             return new int[] {gridX, gridY};
         }
+        */
 
         /**
          * Checks if amount string is present
          * @return true if amount string is present
          */
+        /*
         private boolean checkAmountPresent() {
             String amount = getDetection();
             if (amount.contains("TOTALE"))
@@ -247,6 +254,7 @@ class RawBlock {
             else
                 return false;
         }
+        */
 
         /**
          * Search string in text
