+++ app/src/test/java/com/ing/software/ticketapp/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.ing.software.appscontrini;
+package com.ing.software.ticketapp;
 
 import org.junit.Test;
